refactor: make useSetIsDashboardInEditMode generic to any page layout
