--- packages/twenty-front/src/modules/action-menu/actions/record-actions/single-record/dashboard-actions/types/DashboardSingleRecordActionKeys.ts
@@ -1,5 +1,0 @@
-export enum DashboardSingleRecordActionKeys {
-  EDIT_DASHBOARD = 'edit-dashboard-single-record',
-  SAVE_DASHBOARD = 'save-dashboard-single-record',
-  CANCEL_DASHBOARD_EDITION = 'cancel-dashboard-edition-single-record',
-}
