+++ packages/twenty-front/src/modules/page-layout/actions/PageLayoutSingleRecordActionKeys.ts
@@ -1,0 +1,5 @@
+export enum PageLayoutSingleRecordActionKeys {
+  EDIT_LAYOUT = 'edit-layout-single-record',
+  SAVE_LAYOUT = 'save-layout-single-record',
+  CANCEL_LAYOUT_EDITION = 'cancel-layout-edition-single-record',
+}
