+++ packages/twenty-front/src/modules/page-layout/hooks/useSetIsPageLayoutInEditMode.ts
@@ -1,12 +1,12 @@
-import { DashboardSingleRecordActionKeys } from '@/action-menu/actions/record-actions/single-record/dashboard-actions/types/DashboardSingleRecordActionKeys';
 import { forceRegisteredActionsByKeyState } from '@/action-menu/actions/states/forceRegisteredActionsMapComponentState';
+import { PageLayoutSingleRecordActionKeys } from '@/page-layout/actions/PageLayoutSingleRecordActionKeys';
 import { PageLayoutComponentInstanceContext } from '@/page-layout/states/contexts/PageLayoutComponentInstanceContext';
 import { isPageLayoutInEditModeComponentState } from '@/page-layout/states/isPageLayoutInEditModeComponentState';
 import { useAvailableComponentInstanceIdOrThrow } from '@/ui/utilities/state/component-state/hooks/useAvailableComponentInstanceIdOrThrow';
 import { useRecoilComponentCallbackState } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentCallbackState';
 import { useRecoilCallback } from 'recoil';
 
-export const useSetIsDashboardInEditMode = (pageLayoutIdFromProps: string) => {
+export const useSetIsPageLayoutInEditMode = (pageLayoutIdFromProps: string) => {
   const pageLayoutId = useAvailableComponentInstanceIdOrThrow(
     PageLayoutComponentInstanceContext,
     pageLayoutIdFromProps,
@@ -17,19 +17,19 @@ export const useSetIsDashboardInEditMode = (pageLayoutIdFromProps: string) => {
     pageLayoutId,
   );
 
-  const setIsDashboardInEditMode = useRecoilCallback(
+  const setIsPageLayoutInEditMode = useRecoilCallback(
     ({ set }) =>
       (value: boolean) => {
         set(isPageLayoutInEditModeState, value);
         set(forceRegisteredActionsByKeyState, (prev) => ({
           ...prev,
-          [DashboardSingleRecordActionKeys.EDIT_DASHBOARD]: !value,
-          [DashboardSingleRecordActionKeys.SAVE_DASHBOARD]: value,
-          [DashboardSingleRecordActionKeys.CANCEL_DASHBOARD_EDITION]: value,
+          [PageLayoutSingleRecordActionKeys.EDIT_LAYOUT]: !value,
+          [PageLayoutSingleRecordActionKeys.SAVE_LAYOUT]: value,
+          [PageLayoutSingleRecordActionKeys.CANCEL_LAYOUT_EDITION]: value,
         }));
       },
     [isPageLayoutInEditModeState],
   );
 
-  return { setIsDashboardInEditMode };
+  return { setIsPageLayoutInEditMode };
 };
